findPossibleMoves(),rollDice(),symmetrical init()
rollDice actually determines outcome by throwing loads of dices. would
be nice to make it quicker with taking the info out of some normal
distribution. But for now, that is not determined yet. rollDiceTest is
just a function to test that, but is not needed for the rest of the
project.